CLIENT-374 , CLIENT-1436 Introduce QOpenGLFunctions for all OpenGL calls Replace QOpenGLWindow for QOpenGLWidget
Several different approaches have been tried
to optimize the current approach for zooming/panning
and scrolling but none of them seem to improve the current
one. A more proper refactoring of the rendering and scrolling
needs to be done but at least for now this code
should work in all platforms.

Change-Id: Ifd7d933c2b9d862312b61fefb422f0df603ea1d8