Adding support for Protobuf, Capnproto and Yajl (json).
Two binary formats holding the same data as the current JSON feature format have been created,
one for Protobuf and one for Capnproto.

Support for parsing the current JSON format with the Yajl parser was added.

New programs added:

convert-feature-format:
It converts between the different feature formats (JSON, Protobuf and Capnproto).

feature-fileformat-benchmark:
It performs a benchmark for parsing and serialization of the different formats.

verify-formats.sh:
It verifies that converting back and forth between the different formats doesn't change
the data.

The current state of the implementation contains some local paths to Eriks desktop.
So some modifications need to be done to try this out one some other computer.

Change-Id: I9c14efb6b37c66c05018320fbf29bcfebd59877d